client/pyproject.toml
[project]
name = "messenger-with-crypting"
version = "0.1.0"
description = "Messenger with post-quantum encryption"
readme = "README.md"
authors = [
    { name = "alya san", email = "jiguli.344a@gmail.com" }
]
requires-python = ">=3.13, <3.14"
dependencies = [
    "cryptography",
    "dearpygui",
    "liboqs-python",
    "msgspec[toml]",
    "orjson",
    "pydantic",
    "torch",
    "transformers",
    "websockets",
]

[tool.uv.sources]
liboqs-python = { git = "https://github.com/open-quantum-safe/liboqs-python.git", tag = "0.12.0" }

[dependency-groups]
dev = [
    "ruff",
]

[tool.ruff]
src = ["./src/"]
lint.select = ["ALL"]
lint.ignore = ["D10", "D203", "D212"]
